[l10n] Implement remote locales
This CL changes how locale data is fetch to the same approach that
Lighthouse is using. The difference is, DevTools bundles more than
the en-US locale upfront, so we need to specify that explicitly.

Long term, we should have one central place where all the DevTools
locales are configured. That is, which ones are available, which
ones are shipped and which ones are fetched remotely. For now,
its configured in a few places.

R=tvanderlippe@chromium.org

Bug: chromium:1163928
Change-Id: Iba164eac05bab14a63301c7453f09021ee06991a