Add --override variable generator for dark mode
This CL adds a small helper to the component docs that can take in a
color and generate code to use when migrating that color to an
explicit `--override-*` variable as part of the dark mode migration.
It will use the legacy theme color patching to invert the color and
generate some sample CSS for dark mode.

https://imgur.com/a/O9OrL19

Bug: chromium:1152736
Change-Id: I7c24cdb67577f5c543c734203fe6f5c60635e6ab