[live-edit] Don't attempt to compile modules
When changing resources during live-edit, we send JavaScript files to V8
to compile and show any compile errors in the editor. If the script
being edited is a module, compilation fails with errors like "Cannot use
import statement outside a module". This CL adds tracking for whether
scripts are for modules, and skips compilation when a module is being
edited.

Fixed: chromium:1202495
Change-Id: I15842ad0565c1b67719c7ed212a68e5992550f13