Add invoke to Page.setLifecycleEventsEnabled in ResourceTreeModel
In order to dispatch lifecycle events from the backend the
`lifecycle_events_enabled_` flag in the page agent needs to be set to
true. This is done via the Page.setLifecycleEventsEnabled CDP method.
The ResourceTreeModel currently doesn't have an invoke to the CDP
method and so we add a new method to the class which includes it.

Bug: 1208453
Change-Id: I32ecc3ec680dbfedda2f9381a47ef0725a209ff2