Match extended grid line colors to line color
Extended grid lines are normally light fray, which makes them really
hard to see on some backgrounds.
This change makes extended grid lines use the same color as the lines
themselves, so they can be seen on all backgrounds more easily.

The grid container already has a solid line around it, so it's easy to
see where the grid starts and ends, even if lines and extended lines are
the same color.

Before/after: https://imgur.com/a/OvW46yn

Bug: 1207301
Change-Id: I5d363edf00acf8aa99e1d642c9088f0880286a10