Fix DOM breakpoints not being restored on page reload
this._currentURL() is empty when the page is reloaded because the
new document has not been requested yet and the old one has been
removed. Therefore, we need to request the document and wait for it.
Also, the CL implements proper suspend/resume functionality to disable
event handling when the model is suspended.

Fixed: chromium:824099
Change-Id: I2123ef6f18df18cd28fc4950d9b3d0499a542fd3