Debounce scheduled render calls
If we attempt to schedule renders once one had already been scheduled,
we automatically follow with a second render when the first has
finished. Effectively we mostly debounce, i.e., if you attempt to
schedule 20 calls before the first has started, we'll do 2 renders.

This CL updates the logic to a two-step one. In the first instance, a
render callback is placed in a pending set, such that if we receive any
more schedule calls when the first is pending we ignore them. If the
callback has started, however, we do schedule a follow-up render.

R=jacktfranklin@chromium.org

Bug: 1209820
Change-Id: Ib8fc82d0cabfd10a370f2a1689dfad415ed8494d