Fix leaked mousemove event handler blocking input number steps
I suspect this was caused by some TypeScriptification work a while ago.
The issue is that when we create a local anonymous function to bind to
'mousemove' on targetDocument, we cannot remove this event listener
later, which means we keep capturing this event after the drag. This
creates subtile problems, for example, in the associated bug that
certain UI elements is unclickable. This only happens after any
ResizerWidget is dragged.

Bug: chromium:1155447
Change-Id: I7f042c6b25dd52d1c127fe2f5c68db834242d69c