Add missing DEPS on protocol
When we made the protocol importable, the `ts_library` targets no
longer recompile when we update the protocol. To make sure that
all `ts_library` targets recompile when doing so, we need to
add the proper DEPS on them.

This patch makes sure that https://crrev.com/c/2902705 is correctly
recompiling (and failing DevTools) with the protocol update.

Longterm, we should enforce that all imported modules are listed
in the BUILD.gn file, to make sure that they recompile when required.

R=szuend@chromium.org

Bug: 1210830
Change-Id: I9fa108f84edf4f431328aa26e2df25fb1f440732