Deduplicate watch expression errors
Increased precision in script attribution in the chromium back-end
landed in https://crrev.com/c/2595528 and causes the DevTools front-
end to be able to distintuish the source locations of /different/
evaluations of the same watch expression. This, in turn, causes the
console to not deduplicate error messages, as the differing source
locations between the watch expression evaluations make it plain that
the errors are comming from different scripts. The user expectation,
however, is that errors from consecutive evaluations of watch
expressions are deduplicated.

This CL deduplicates errors again by supplying a sourceURL to the
watch expression, and using that sourceURL to deduplicate errors
even if the scriptId doesn't match.

Bug: chromium:1211255
Change-Id: Ia42ae3b2424072c4418b73008307b9e4b76674b5