[DataGrid] fix virtualisation & scrollbar height
This CL fixes a bug where the height of the data-grid scrollbar would
change drastically as you scrolled. This was a bug in the virtual
rendering logic; the bottom "filler" row that we increase/decrease the
height of as you scroll (to emulate the hidden rows that we aren't
actually rendering) was being calculated based on the number of
renderable rows; however that variable refers to the rows we are
actually rendering, not actually the amount of rows in total. Swapping
it to use all the rows that are in the table (minus those that are
filtered out because e.g. the user has searched for some text) fixes
this, ensuring the scrollbar remains the same size as you scroll.

Bug: 1211346
Change-Id: Ie91d32afd99bdc1ee624834e1cc0dd889b7c6031