Adding role=group to InfoBars that have actions assoicated with them
On Windows, with the Narrator screen reader, the "pretty print" infobar
displayed in the Sources panel is announced as a group. This is good
because it's a section of devtools that contains a set of UI elements
but not a major part of the interface hierarchy.
However on Linux, with the Orca screen reader, this infobar isn't
announced as a group.

This CL adds the role=group explicitely to infobars that have actions.

Bug: 1211338


Merge remote-tracking branch 'origin/master'


Merge remote-tracking branch 'origin/master'


Match extended grid line colors to line color

Extended grid lines are normally light fray, which makes them really
hard to see on some backgrounds.
This change makes extended grid lines use the same color as the lines
themselves, so they can be seen on all backgrounds more easily.

The grid container already has a solid line around it, so it's easy to
see where the grid starts and ends, even if lines and extended lines are
the same color.

Bug: 1207301
Change-Id: I308568d52c57c706d5ae0cc07245c6c9deb08267