Sources: Add alert to pretty-print info bar learn more message.
When the user clicks "Learn More" in the pretty-print info-bar in
sources, a description is added to the screen describing how the
pretty-print functionality works. However, nothing is announced
to screen readers. This PR marks that info as an alert, so that
when it is revealed, Screenreaders announce the info.

Change-Id: Iaaa56623c9b98ffacce34d57c38721870dceb15a