[ProtocolMonitor] highlight sent messages
This CL adds back the feature where outgoing messages are slightly
highlighted in the protocol monitor to enable them to be
distinguished. To do this I've added support for custom CSS variables
to the DataGrid that can be overriden via a `styles` property on an
individual row. This way we avoid having to work around shadow roots
and such like to pass styling through, and we limit how much
customisation people can do when rendering with custom colors.

Bug: 1211346
Change-Id: I280028bb5706065078b79f3a424d572391b97254