Skip VSCode settings update on error
I had a rogue comma in my settings.json file. This prevented the require
from running correctly, which in turn stopped gclient sync from running.
This CL adds a try-catch around the code so that on any error, the
update is skipped.

This CL also adds a bypass option, so that people can opt out of syncing
their VSCode settings entirely.

R=jacktfranklin@chromium.org

Bug: None
Change-Id: I3d90dc0546fcc145c881cb1aa49964e79db3f24f