[Dark Mode] ban enableLegacyPatching: true
This CL sets the type for `enableLegacyPatching` to `false` rather
than `boolean`, such that we ensure no legacy code and no new code can
make use of this behaviour.

In a future CL we will remove the legacy patching entirely, but doing
it like this means it's a smaller, easier CL to revert should we find
any big issues.

Bug: chromium:1152736
Change-Id: I7c411683c995090231520ae70324cced70094d44