Fix formatting for class definitions
The version of TypeScript we use now properly emits class field
definitions. Since this is an ES2022 feature, we need to bump
the version of Acorn to support parsing class field definitions.
Additionally, we need to update the formatter to handle the
spacing around the tokens in a class field definition.

We can revert the local patch to Acorn once
https://github.com/acornjs/acorn/pull/1041 is merged. I will
also work on updating the ESTree type definitions.

R=bmeurer@chromium.org

Bug: none
Change-Id: Id2232e6ce2d8a0a35002e87f181259bb60d26feb