Add issue codes and draft descriptions for Attribution Reporting
This CL adds draft descriptions for Attribution Reporting API issues.
Note that the issues are not shown yet and any AR issue arriving
via CDP is currently ignored.

AR issues will be enabled when the UI side is finalized.
Landing draft descriptions early gives more time for DevRel to
iterate on the descriptions.

R=sigurds@chromium.org

Bug: chromium:1190735
Change-Id: I7e4cd18038cdebe5aebd6f07671f84986a3230d2