Don't use targetId/frameId correspondence in DOMModel
This CL is a test balloon to see if we can get away with not using
the fact that targetId and main frameId of the target correspond.
We assume that the document is only requested in a frame after the
main frame of the target became known to DevTools. If this is not
the case, then this CL will cause problems.

Bug: chromium:1213943
Change-Id: If3526508e9442e7289d9ca744af17ab641ab594c