Reland: [i18n] Downstream latest LH i18n library changes
This CL is a straight-up reland of https://crrev.com/c/2905493.
The CL was reverted as it increased the frequency of flakes on the
Leak detection bot. The underlying issue was fixed and now this CL
can be landed.

The reland had to resolve merge conflicts and as such was not created
via Gerrit.

Original description:

This CL upgrades the i18n LH library from revision
f3d0e3459d8fd15b055148dec0ae4e430df6495b to revision
15dad3944c2754a9e147fef9ef99934169e2ca08.

The LH changes include a major re-write on how ICU messages are
handled, namely a move from string IDs to IcuMessage objects.
The changes were mostly adapted as-is and the i18n-bundle.d.ts
updated to reflect the renamed functions.

To make getFormatLocalizedString work, we had to weaken a runtime
type check in `isIcuMessage`. DevTools allows for complex objects
as placeholders, not just numbers and strings.

Fixed: chromium:1208456
Change-Id: Iebe0324880eb3de1f0b7bb2ea0ef9fcc72272883