[i18n] Add new class for caching localized strings and formatters
This CL adds a new `LocalizedStringSet` class that serves as a file
local cache for translations. We change the return type
of `registerUIStrings` to return this new class instead, which helps
later to greatly speed up DevTools localization.

On a test page that only created console log messages, the self time
of `getLocalizedString` was reduced from ~400ms to ~20ms.

Drive-by: This change also removes the lookup handling of ModuleUIStrings
as they no longer exist.

Bug: chromium:1203089
Change-Id: I8a20228ef8ec9e355a0523f01a2cbe9bb5a3912c