[l10n] Use fallback locale when fetching remote locale fails
Currently, when fetching the remote locale fails, then the 404 exception
races with installing the en-US fallback locale. This CL fixes that by
first loading the bundled en-US and then attempting to fetch the
possible remote locale.

This CL also introduces a new method on "DevToolsLocale" that allows
us to force the language back to english when fetching the preferred
locale fails.

R=aerotwist@chromium.org

Bug: chromium:1163928
Change-Id: I3499d7c0489aafcea53f962e878d31f1fcb431ac