Fix the visibility of soft context menu item in windows high-contrast
Soft context menu items were not visible in high-contrast mode because
the force-color-adjust property was set to none and that meant the items
defaulted back to the main body color, which didn't work when the
background of the menu was dark.

In this change, I'm using the standard CanvasText color that works well
in any of the high-contrast themes.

I also discovered that menu items had a top and bottom transparent
border that got changed to the color of the background on hover. So
essentially, the borders were never visible, they always were either
transparent or blended with the background color.
So I replaced this with padding space instead. Otherwise, the force
color mode transformed those transparent borders in actually visible
colors which didn't look good.

The last thing I had to fix was on hover, windows was added a backplate
to the item's label which actually made the text unreadable in the white
high-contrast mode. So I had to use force-color-adjust:none in this case
but still kept the Highlight and HighlightText colors for sufficient
contrast.

Before/After: https://imgur.com/a/UyNtPZV

Bug: 1217091
Change-Id: I65cada2a8501ddeb9c755a02652797700010594a