Refactor request-resolving logic to RequestResolver
When resolving requestIds to requests, it can happen that the request
has not yet arrived from the back-end. We currently have logic in the
issues panel that takes care of such cases by registering a listener
on the network log to learn about a request with a certain requestId
when it becomes known to the DevTools front-end. This CL refactors that
logic into its own class, and adds a promisified interface as well.

Bug: chromium:1217975
Change-Id: Ide7f6d41911f3d19fa392ed94230bd24cb48c1d4