Consolidate request displaying in issues
We previously had several pieces of code that displayed a request, now
all use AffectedResourcesView#createRequestCell.

This CL changes behavior for MixedContentIssues slightly: We only
display the dedicated mixed content resources view (instead of also
displaying the requests view, which has the same information with
less detail). Additionally, we now display requests for which we
do not have a network request. This triggers a small quirk where
we display some requests twice if the renderer issues a prefetch/
cached fetch.

Bug: chromium:1217975
Change-Id: Ice8eb0298f81c7bc1e98800cfe5c1ab5b462ff89