[recorder] Ensure uniqueness of selectors
This CL changes the step format to support a list of
selectors that is needed to query elements inside Shadow DOM.
Also it adds a verification code that checks that a selector
chain resolves to a single element. Note that for CSS selectors
we don't perform the verification because we assume the generated
CSS selector is unique.

Fixed: chromium:1215440
Change-Id: Iecd5377970ac5371263db3e20f786e1ffaffbe66