Add rule to enforce deep equality comparisons for array assertions
Not every chai assertion uses deep comparisons. For arrays, we need
to make sure that not just the array itself, but also its contents
and its members are equal. There are only a subset of assertions
in chai that do that, so allow only these in our test assertions.

R=dsv@chromium.org

Fixed: 1218015
Change-Id: Ib7eb5864cf65a54ef17b9d637cd8e3b94323edb6