[issues] Add details view for Attribution Reporting issue
This CL adds the AttributionReportingIssueDetailsView that shows
affected resources for Attribution Reporting API issues. This CL
also adds the table for the missing "attribution-reporting"
Permissions Policy issue. The issue can happen either during an
attribution source event (aka an impression) or an attribution
(aka a conversion). Depending on the event, either the network
request or the source HTML element is available.

Screenshots: https://imgur.com/a/JrwgrDd

R=sigurds@chromium.org

Bug: chromium:1190735
Change-Id: I29051af4c041878624bccba4544ec840a02caeb1