Fix misplaced "==$0" in the Elements tree
This is caused by the ElementsTreeElement creating a separate <span>
and "outsourcing" the formatting to our JS highlighter, which somehow
causes the `.selected-hint` element (holder of "==$0"), to be misplaced.
This CL puts the `.selected-hint` element back into document flow
when it's displayed, so that it won't be misplaced on some other element.

Screenshot: https://imgur.com/a/ph5Zn74

Fixed: chromium:1087173
Change-Id: If34652bd4db002fbc54c6c99f04ea619ee8e5049