Add component for linking to network requests
This CL adds a component that can be used to link to a network request.
The component optionally uses a RequestResolver to resolve the request
if only a requestId (in the form of an affected request) is provided.
The component updates its state once the request becomes available,
thereby supporting plasticity of the interface.

Bug: chromium:1217975
Change-Id: I94faa598bb520d4cbdf003ca7858ec759fcc8a6c