Update ConsoleTestRunner.visibleIndices to handle fractional layouts
In an upcoming change for Blink (WIP cl/2824938), we're trying to land
a change that removes line-height clamping from layout. This results in
lines having additional precision when laying out. For devtools, the
console output lines are styled as font-size: 12px, line-height 1.2em,
which results in fractional pixel measurements.

This CL adjusts how console visible indices are calculated for testing.
ConsoleViewport itself uses getBoundingClientRect for the cumulative
heights, but this can result in subpixel precision differences when
manually calculating visible indices.

Validated that the two web_test callers of this method pass as-is, as
well as with the WIP change applied.
http/tests/devtools/console/console-viewport-control.js
http/tests/devtools/console/console-viewport-indices.js

Bug: 1023654
Change-Id: Iba1799e36057be76affc628c5157e7e43de3ad98