[settings] Add setting to disable auto-focus on the sources panel
This adds a setting to DevTools to disable auto-focusing on the
sources panel if a Debugger.paused event is received.
The context here is that extensions may be attached that do not
want that behavior (e.g. if source debugging is happening in
another panel that is provided by the extension).

Bug: b/182581017
Change-Id: I7a50c0c6eb7d5eb7abe45bcfafc40e394a4b8e17