Use CSSStyleSheetHeader.isConstructedNonModule instead of .isConstructed
A CSSStyleSheet imported as a CSS module script has the constructed
flag set
(https://wicg.github.io/construct-stylesheets/#cssstylesheet-constructed-flag).
This is because it is useful to treat imported CSSStyleSheets the same
way as stylesheets created via `new`. For example we want to be able
to add both to document.adoptedStyleSheets, and we want both to treat
@import rules the same way.

However, we want DevTools to display them differently. Sheets
created with `new` don't have a source file to show, but imported
sheets do, so we want the latter to be shown more like "normal"
stylesheets.

So, this change switches from isConstructed to isConstructedNonModule,
added in [1], so that we still link to source files for 'constructed'
stylesheets imported as CSS module scripts.

[1] Blink change: https://chromium-review.googlesource.com/c/chromium/src/+/2964539

Bug: 1125357
Change-Id: I7163966c66c990647e40a72d0ef9d6ba30f41cfb