Improve Console display of strings
We recently changed how we display the output to ensure that we always
display valid JavaScript literals. Unfortunately, the current solution
can be confusing to users, who might feel that the amount of escaping is
excessive and makes the output unreadable.

This change introduces `formatAsJSLiteral` that given a string escapes
appropriate special characters and wraps the string in either single
quotes, double quotes, or back ticks depending on the string content.
The function furthermore escapes script tags to ensure that the result
can be safely copied to HTML.

Both regular Console output and preview have been updated to use this
formatter instead of JSON.stringify for displaying strings.

This change adds unit tests for the new formatter and rebaselines all
e2e tests that depend on Console outputs.

Doc: https://goo.gle/devtools-strings
Bug: 1208389
Change-Id: Iebfe39c2e40c21aa5a813a1ebd26a5109fa59ab6