[Recorder] implement buffering for keyboard interactions with input elements
With this CL, when the user focuses on an input element, the client
enters a special mode that handles event slightly differently:

- it publishes keyup events from the previous keyboard navigation
immediately.
- it captures the rest of keyboard events into a buffer.
- when a change or blur event is sent to the element, it interprets
the buffer and correctly publishes keyboard events that lead
to the navigation.

This way we lose some precise events that happen during interaction
with an input as we capture only the final result but we are also
able to maintain keyboard navigation events and handle complex
inputs like input[type=color|slider] where keyboard or mouse events
are not enough to capture the input.

Fixed: chromium:1221037, chromium:1221046
Change-Id: Id2949c1ec88c5becbffd262e5b95628d6e6300f4