[Recorder] Recorder and replay click event offsets
This CL adds offsetX/offsetY to click events allow to be more
precises about where an element is clicked. This solves a problem
of clicking on complex inputs such as date inputs.

As a drive-by I fixed the wrong viewport measurements being captured.
Happy to extract to a separate CL if the clicking part will be
controversial.

Bug: chromium:1221043
Change-Id: I9e18a3e77ddb99f93712f1065da1a0fe3a1cbfa7